refactor(primitives): promote BlockRef to a top-level type
