+++ packages/zaino-mempool-service/src/tests.rs
@@ -20,7 +20,7 @@ use zaino_source::{
 };
 
 use zaino_mempool::config::MempoolConfig;
-use zaino_mempool::ports::BlockRef;
+use zaino_primitives::types::BlockRef;
 
 #[cfg(feature = "tip_aware_mempool")]
 use zaino_mempool::ports::{NfsEpochObserver, NonFinalizedEpoch};
