+++ packages/zaino-mempool/src/ports.rs
@@ -20,32 +20,10 @@
 use std::sync::Arc;
 
 use tokio::sync::broadcast;
-use zaino_primitives::types::{BlockHash, Height};
+use zaino_primitives::types::BlockRef;
 
 use crate::snapshot::MempoolSnapshot;
 use crate::update::MempoolUpdate;
-
-/// A minimal `(hash, height)` reference to a block.
-///
-/// A named pair rather than a tuple because the mempool compares these for
-/// equality in the one place that decides whether a transaction set is still
-/// coherent with the chain; `tip.hash != other.hash` reads correctly where
-/// `tip.0 != other.0` invites the wrong field.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct BlockRef {
-    /// The block's hash.
-    pub hash: BlockHash,
-    /// The block's height.
-    pub height: Height,
-}
-
-impl BlockRef {
-    /// Build a reference from the pair `zaino-source`'s tip ports answer with.
-    pub fn from_tip(tip: (BlockHash, Height)) -> Self {
-        let (hash, height) = tip;
-        Self { hash, height }
-    }
-}
 
 /// Every validator question the tip-agnostic mempool core asks.
 ///
