+++ packages/zaino-primitives/src/types/rpc.rs
@@ -48,7 +48,7 @@ mod spent_info;
 mod subtree_roots;
 mod tx_out;
 
-pub use address_deltas::{AddressDeltas, AddressDeltasRequest, BlockRef};
+pub use address_deltas::{AddressDeltas, AddressDeltasRequest};
 pub use block_deltas::{BlockDelta, BlockDeltas, InputDelta, OutputDelta};
 pub use block_header::BlockHeaderVerbose;
 pub use block_subsidy::{BlockSubsidy, FundingStream, LockboxStream};
